% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIndex.R
\name{createIndex}
\alias{createIndex}
\title{Create Index file}
\usage{
createIndex(file, filePath)
}
\arguments{
\item{file}{File name}

\item{filePath}{Folder path for file to be written to}
}
\value{
None
}
\description{
Creates an index file to loccate which fst file a particular position is
located in
}
