% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPatientFile.R
\name{createPatientFile}
\alias{createPatientFile}
\title{Create Patient File}
\usage{
createPatientFile()
}
\value{
None
}
\description{
Binds all patient of a certain cancer into one large data file
}
