% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileNames.R
\name{enterVCFFileName}
\alias{enterVCFFileName}
\title{Reading files}
\usage{
enterVCFFileName()
}
\value{
None
}
\description{
Reads in VCF and Depth files
}
\examples{
enterVCFFileName()
enterDepthFileName()
}
