% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSampleMerge.R
\name{multiSampleMerge}
\alias{multiSampleMerge}
\title{Multi sample merge}
\usage{
multiSampleMerge()
}
\value{
None
}
\description{
Merges multiple samples from single patient based on preference order of
samples
}
