% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFiles.R
\name{processVCF}
\alias{processVCF}
\title{Process files}
\usage{
processVCF(vcf_File)
}
\arguments{
\item{vcf_File}{VCF file to be processed}

\item{depth_File}{DEPTH file to be processed}
}
\description{
Processes and parses VCCF and Depth files
}
\examples{
processVCF(xxx.vcf)
processDepth(xxx.depth)

}
