% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMerge.R
\name{sampleMerge}
\alias{sampleMerge}
\title{Merging a signle sample}
\usage{
sampleMerge()
}
\arguments{
\item{filePath}{folder path for final data}
}
\value{
None
}
\description{
Processes and merges VCF and Depth of single sample
}
